[MERPWH-795] Get module version by sql-query for 'res.company' model (#177)
* [FIX][MERPWH-795] Computation of the 'base_version'

* [FIX][MERPWH-795] Get module version by sql-query for 'res.company' model

* [IMP][MERPWH-795] Computation 'ventor base' bodule version by sql-query

* [ADD][MERPWH-795] Readme file
