+++ outgoing_routing/__manifest__.py
@@ -3,7 +3,7 @@
 
 {
     'name': 'Picking and Reservation Strategy',
-    "version": "14.0.1.0.0",
+    "version": "14.0.1.0.1",
     'author': 'VentorTech',
     'website': 'https://ventor.tech/',
     'license': 'LGPL-3',
